move rendering to separate plugin
